inputan tarif layanan

+++ app/Http/Controllers/TarifController.php
@@ -152,11 +152,68 @@ class TarifController extends APIController
             return $this->sendError($th->getMessage(), $th->getCode());
         }
     }
+    function update_tarif_pasien(Request $request)
+    {
+        try {
+            $validator = Validator::make(request()->all(), [
+                'kodebooking' => 'required',
+                'antrian_id' => 'required',
+                'kunjungan_id' => 'required',
+                'kodekunjungan' => 'required',
+                'layanan' => 'required',
+                'jumlah' => 'required',
+                'harga' => 'required',
+                'diskon' => 'required',
+            ]);
+            if ($validator->fails()) {
+                return $this->sendError($validator->errors()->first(), 400);
+            }
+            // add layanan
+            if ($request->layanan) {
+                $tarif = Tarif::find($request->layanan);
+                $user = Auth::user()->id;
+                if ($user ==  $tarif->user) {
+                    $layanans = Layanan::updateOrCreate(
+                        [
+                            'kodebooking' => $request->kodebooking,
+                            'antrian_id' => $request->antrian_id,
+                            'kodekunjungan' => $request->kodekunjungan,
+                            'kunjungan_id' => $request->kunjungan_id,
+                            'tarif_id' =>  $tarif->id,
+                        ],
+                        [
+                            'nama' => $tarif->nama,
+                            'harga' => $request->harga,
+                            'jumlah' =>  $request->jumlah,
+                            'diskon' => $request->diskon,
+                            'subtotal' => ($request->harga * $request->jumlah) - ($request->harga * $request->jumlah * $request->diskon / 100),
+                            'klasifikasi' => $tarif->klasifikasi,
+                            'jaminan' => $request->jaminan,
+                            'pic' => Auth::user()->name,
+                            'user' => Auth::user()->id,
+                            'tgl_input' => now('Asia/Jakarta'),
+                        ]
+                    );
+                    return $this->sendResponse($layanans, 200);
+                } else {
+                    return $this->sendError('Tidak bisa diedit oleh orang lain', 401);
+                }
+            }
+        } catch (\Throwable $th) {
+            //throw $th;
+            return $this->sendError($th->getMessage(), $th->getCode());
+        }
+    }
     function delete_tarif_pasien(Request $request)
     {
+        $user = Auth::user()->id;
         $tarif = Layanan::find($request->tarif);
-        $tarif->delete();
-        return $this->sendResponse('Ok', 200);
+        if ($user ==  $tarif->user) {
+            $tarif->delete();
+            return $this->sendResponse('Ok', 200);
+        } else {
+            return $this->sendError('Tidak bisa dihapus oleh orang lain', 401);
+        }
     }
     function get_layanan_kunjungan(Request $request)
     {
